fix warnings

+++ interface/src/avatar/Avatar.cpp
@@ -585,7 +585,7 @@ bool Avatar::shouldRenderHead(const RenderArgs* renderArgs) const {
 
 // virtual
 void Avatar::simulateAttachments(float deltaTime) {
-    for (int i = 0; i < _attachmentModels.size(); i++) {
+    for (int i = 0; i < (int)_attachmentModels.size(); i++) {
         const AttachmentData& attachment = _attachmentData.at(i);
         auto& model = _attachmentModels.at(i);
         int jointIndex = getJointIndex(attachment.jointName);
@@ -949,14 +949,14 @@ void Avatar::setAttachmentData(const QVector<AttachmentData>& attachmentData) {
     AvatarData::setAttachmentData(attachmentData);
 
     // if number of attachments has been reduced, remove excess models.
-    while (_attachmentModels.size() > attachmentData.size()) {
+    while ((int)_attachmentModels.size() > attachmentData.size()) {
         auto attachmentModel = _attachmentModels.back();
         _attachmentModels.pop_back();
         _attachmentsToRemove.push_back(attachmentModel);
     }
 
     for (int i = 0; i < attachmentData.size(); i++) {
-        if (i == _attachmentModels.size()) {
+        if (i == (int)_attachmentModels.size()) {
             // if number of attachments has been increased, we need to allocate a new model
             _attachmentModels.push_back(allocateAttachmentModel(attachmentData[i].isSoft, _skeletonModel.getRig()));
         }
